FEAT- 댓글 도메인 리팩토링

+++ src/main/java/es/board/repository/entity/repository/infrastructure/PostTestRepository.java
@@ -30,7 +30,7 @@ public class PostTestRepository implements PostRepository {
 
 
     @Override
-    public PostEntity findPostDetail(long id) {
+    public PostEntity findPostDetail(int id) {
         return null;
     }
 }
